move to embed to CalendarEvent type
